Transaction field "amount" variable type changed to double from int. RMP-5227 updated.

+++ app/src/main/java/suncor/com/android/model/pap/transaction/Transaction.java
@@ -213,7 +213,7 @@ public class Transaction {
     }
 
     public static class PetroRedeemedRewards{
-        public int amount;
+        public double amount;
         public String description;
     }
